Added user input

+++ ConsoleApp1/CalCulator.cs
@@ -7,18 +7,18 @@ namespace ConsoleApp1
     public class CalCulator
     {
         // Add two integers and returns the sum  
-        public int Add(int num1, int num2)
+        public double Add(double num1, double num2)
         {
             return num1 + num2;
         }
 
         // Multiply two integers and retuns the result  
-        public int Multiply(int num1, int num2)
+        public double Multiply(double num1, double num2)
         {
             return num1 * num2;
         }
         // Subtracts small number from big number  
-        public int Subtract(int num1, int num2)
+        public double Subtract(double num1, double num2)
         {
             if (num1 > num2)
             {
@@ -29,7 +29,7 @@ namespace ConsoleApp1
 
         }
         //performing Division on two float variables.  
-        public float Division(float num1, float num2)
+        public double Division(double num1, double num2)
         {
             return num1 / num2;
         }
